+++ crates/json-abi/tests/it/main.rs
@@ -1,0 +1,9 @@
+#[macro_use]
+mod macros;
+
+mod abi_items;
+mod abis;
+mod event_params;
+mod params;
+mod state_mutability;
+mod test;
