+++ crates/json-abi/tests/it/params.rs
@@ -1,5 +1,3 @@
-mod test_helpers;
-
 use alloy_json_abi::{InternalType::Struct, Param};
 
 #[test]
