+++ crates/json-abi/tests/it/state_mutability.rs
@@ -1,5 +1,3 @@
-mod test_helpers;
-
 use alloy_json_abi::StateMutability;
 
 #[test]
