+++ crates/sol-types/tests/macros/main.rs
@@ -1,0 +1,2 @@
+mod sol;
+mod udvt;
