+++ crates/syn-solidity/src/item/function.rs
@@ -403,11 +403,31 @@ impl FunctionBody {
 mod tests {
     use super::*;
     use pretty_assertions::assert_eq;
+    use proc_macro2::Span;
     use std::{
         error::Error,
         io::Write,
         process::{Command, Stdio},
     };
+    use syn::parse_quote;
+
+    #[test]
+    fn modifiers() {
+        let none: ItemFunction = parse_quote! {
+            modifier noParens {
+                _;
+            }
+        };
+        let some: ItemFunction = parse_quote! {
+            modifier withParens() {
+                _;
+            }
+        };
+        assert_eq!(none.kind, FunctionKind::new_modifier(Span::call_site()));
+        assert_eq!(none.kind, some.kind);
+        assert_eq!(none.paren_token, None);
+        assert_eq!(some.paren_token, Some(Default::default()));
+    }
 
     #[test]
     #[cfg_attr(miri, ignore = "takes too long")]
